packer: linux: Setup module cache
Setup a cache of the zephyr modules to reduce init time.

Since it appears that writes are much faster than reads we remove the
checkout of the files in the repos and just keep the .git dirs around.
We'll have the run_ci.sh checkout the files for the exact SHAs that are
needed.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>